+++ src/ReqRes/WikiUpdateRequest.cs
@@ -115,6 +115,6 @@ namespace Gloobster.ReqRes
         public string language { get; set; }
         public string link { get; set; }
         public string articleId { get; set; }
-        public decimal rating { get; set; }
+        public double rating { get; set; }
     }
 }
